refactor( CartItem ): change type of quantity to Int from Long

+++ src/main/kotlin/ecommerce/dto/CartItemRequest.kt
@@ -6,5 +6,5 @@ class CartItemRequest(
     @field:Min(1)
     val productId: Long,
     @field:Min(1)
-    val quantity: Long,
+    val quantity: Int,
 )
